--- a/packages/ui/src/council/hooks/useCouncilRemainingPeriod.ts
+++ b/packages/ui/src/council/hooks/useCouncilRemainingPeriod.ts
@@ -1,5 +1,5 @@
 import { sum } from 'lodash'
-import { map, combineLatest, concatMap, EMPTY, merge, Observable, of } from 'rxjs'
+import { map, combineLatest } from 'rxjs'
 
 import { useApi } from '@/api/hooks/useApi'
 import { useObservable } from '@/common/hooks/useObservable'