--- a/packages/ui/src/forum/queries/forum.graphql
+++ b/packages/ui/src/forum/queries/forum.graphql
@@ -164,7 +164,7 @@ fragment ForumPostWithoutReplyFields on ForumPost {
   }
   threadId
   thread {
-     categoryId
+    categoryId
   }
 }