--- a/packages/ui/test/bounty/modals/SubmitJudgementModal.test.tsx
+++ b/packages/ui/test/bounty/modals/SubmitJudgementModal.test.tsx
@@ -33,8 +33,6 @@ import {
 
 configure({ testIdAttribute: 'id' })
 
-
-
 jest.mock('@/common/components/CKEditor', () => ({
   CKEditor: (props: CKEditorProps) => mockCKEditor(props),
 }))