--- a/packages/ui/test/membership/modals/BuyMembershipModal.test.tsx
+++ b/packages/ui/test/membership/modals/BuyMembershipModal.test.tsx
@@ -31,8 +31,6 @@ import { mockUseModalCall } from '../../setup'
 
 configure({ testIdAttribute: 'id' })
 
-
-
 describe('UI: BuyMembershipModal', () => {
   const api = stubApi()
   let transaction: any
@@ -69,7 +67,7 @@ describe('UI: BuyMembershipModal', () => {
       fireEvent.click(screen.getByTestId('email-tile'))
       fireEvent.change(screen.getByTestId('email-input'), { target: { value: 'invalid email' } })
     })
-    expect(submitButton).toBeDisabled()
+    Expect(submitButton).toBeDisabled()
   })
 
   it('Enables button when valid form', async () => {